@@ -3,29 +3,37 @@
   "version": "1.0.0",
   "repository": "https://github.com/bycycle/bycycle.webui",
   "license": "GPL-3.0",
-  "devDependencies": {
-    "autoprefixer": "^9.7.2",
+  "dependencies": {
+    "@babel/core": "^7.7.7",
+    "@babel/plugin-transform-runtime": "^7.7.6",
+    "@babel/preset-env": "^7.7.7",
+    "@babel/runtime": "^7.7.7",
+    "autoprefixer": "^9.7.3",
+    "core-js": "^3.6.1",
     "dotenv": "^8.2.0",
+    "hammerjs": "^2.0.8",
     "npm-run-all": "^4.1.5",
     "ol": "^6.1.1",
-    "postcss": "^7.0.23",
+    "page": "^1.11.5",
+    "postcss": "^7.0.26",
     "prettier": "^1.19.1",
-    "rollup": "^1.27.0",
+    "qs": "^6.9.1",
+    "rollup": "^1.27.14",
+    "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-dotenv": "^0.2.0",
     "rollup-plugin-livereload": "^1.0.4",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-svelte": "^5.1.1",
-    "rollup-plugin-terser": "^5.1.2",
-    "rxjs": "^6.5.3",
-    "sass": "^1.23.6",
-    "sirv-cli": "^0.4.5",
-    "svelte": "^3.14.1"
+    "rollup-plugin-terser": "^5.1.3",
+    "rxjs": "^6.5.4",
+    "sass": "^1.24.0",
+    "svelte": "^3.16.7"
   },
   "scripts": {
     "start": "runcommand dev-server"
   },
-  "browsers": [
+  "browserslist": [
     "> 1%",
     "last 2 versions",
     "firefox esr",
