@@ -1,1 +1,7 @@
-export const INPUT_DEBOUNCE_TIME = 750;
+export const INPUT_DEBOUNCE_TIME = 1000;
+
+// Breakpoints
+export const SM_WIDTH = 576;
+export const MD_WIDTH = 768;
+export const LG_WIDTH = 992;
+export const XL_WIDTH = 1200;
