@@ -1,102 +1,81 @@
+<svelte:options immutable={true} />
+
 <script>
-    import { onMount } from 'svelte';
-    import { fade } from 'svelte/transition';
-    import {
-        directionsFromTerm,
-        directionsFromPoint,
-        directionsToTerm,
-        directionsToPoint,
-        directionsError
-    } from '../directions/stores';
-    import { getDirections } from '../directions';
-    import { submitQuery } from '../query';
-    import { queryTerm, queryPoint, queryResults, queryError } from '../query/stores';
-    import { currentFunction, switchToDirections, switchToQuery } from '../stores';
-    import { contextMenuStore, locationStore } from './stores';
+    import { getContext } from 'svelte';
+    import { setCurrentRoute } from '../routes';
     import { displayLatLong } from './util';
 
-    export let map;
-
+    const map = getContext('map');
     const osmBaseUrl = '//www.openstreetmap.org/#map=17/';
-    const style = { display: 'none', top: 0, right: 0, bottom: 0, left: 0 };
 
+    export let open = false;
+    export let pixel = [0, 0];
+
+    let element;
     let osmUrl = osmBaseUrl;
+    let style;
 
     $: {
-        const [mapWidth, mapHeight] = map.getSize();
-        const { open, x, y } = $contextMenuStore;
+        if (element && open) {
+            const [x, y] = pixel;
+            const [mapWidth, mapHeight] = map.getSize();
+            let top, right, bottom, left;
 
-        style.display = open ? 'block' : 'none';
+            for (const anchor of element.querySelectorAll('li > a')) {
+                anchor.classList.add('map-context-menu-item');
+            }
 
-        if (open) {
-            osmUrl = `${osmBaseUrl}${displayLatLong(getCoordinate(), '/')}`;
+            osmUrl = `${osmBaseUrl}${getCoordinate(true, '/')}`;
+
+            if (x < mapWidth / 2) {
+                left = `${x}px`;
+                right = 'auto';
+            } else {
+                left = 'auto';
+                right = `${mapWidth - x}px`;
+            }
+
+            if (y < mapHeight / 2) {
+                top = `${y}px`;
+                bottom = 'auto';
+            } else {
+                top = 'auto';
+                bottom = `${mapHeight - y}px`;
+            }
+
+            style = `top: ${top}; right: ${right}; bottom: ${bottom}; left: ${left};`;
         }
+    }
 
-        if (x < mapWidth / 2) {
-            style.left = `${x}px`;
-            style.right = 'auto';
-        } else {
-            style.left = 'auto';
-            style.right = `${mapWidth - x}px`;
-        }
+    function close() {
+        open = false;
+    }
 
-        if (y < mapHeight / 2) {
-            style.top = `${y}px`;
-            style.bottom = 'auto';
-        } else {
-            style.top = 'auto';
-            style.bottom = `${mapHeight - y}px`;
+    function handleClick (event) {
+        const target = event.target;
+        if (target.tagName === 'A') {
+            const href = target.getAttribute('href');
+            if (href.charAt(0) === '#') {
+                event.preventDefault();
+            }
         }
+        event.stopPropagation();
+        close();
     }
 
     function handleWhatIsHere() {
-        const coordinate = getCoordinate();
-        const latLong = displayLatLong(coordinate);
-        const args = { term: latLong, point: coordinate, myLocation: $locationStore };
-
-        if ($currentFunction !== 'QUERY') {
-            switchToQuery();
-        }
-
-        $queryTerm = latLong;
-        $queryPoint = coordinate;
-        $queryError = null;
-
-        // XXX: Setting this here keeps results from updating after the
-        //      first time this function is called. I have no idea why.
-        // $queryResults = [];
-
-        submitQuery(args, map);
+        const coordinate = getCoordinate(true);
+        setCurrentRoute('search', { term: coordinate });
     }
 
     function handleGetDirectionsFrom() {
-        const coordinate = getCoordinate();
-        const latLong = displayLatLong(coordinate);
-        const args = {
-            fromTerm: latLong,
-            fromPoint: coordinate,
-            toTerm: $directionsToTerm || 'My Location',
-            toPoint: $directionsToPoint,
-            myLocation: $locationStore
-        };
-        if ($currentFunction !== 'DIRECTIONS') {
-            switchToDirections();
-        }
-        getDirections(args, map);
+        const coordinate = getCoordinate(true);
+        setCurrentRoute('directions', { fromTerm: coordinate, toTerm: '?' });
     }
 
     function handleGetDirectionsTo() {
-        const coordinate = getCoordinate();
-        const latLong = displayLatLong(coordinate);
-        const args = {
-            fromTerm: $directionsFromTerm || 'My Location',
-            fromPoint: $directionsFromPoint,
-            toTerm: latLong,
-            toPoint: coordinate,
-            myLocation: $locationStore
-        };
-        switchToDirections(args);
-        getDirections(args, map);
+        const coordinate = getCoordinate(true);
+        setCurrentRoute('directions', { fromTerm: '?', toTerm: coordinate });
     }
 
     function handleCenterMapHere() {
@@ -108,23 +87,34 @@
         map.zoomToStreetLevel();
     }
 
-    function getCoordinate() {
-        return map.getCoordinateFromPixel([$contextMenuStore.x, $contextMenuStore.y]);
+    function getCoordinate(display = false, sep = ',') {
+        const coordinate = map.getCoordinateFromPixel(pixel);
+        if (display) {
+            return displayLatLong(coordinate, sep);
+        }
+        return coordinate;
     }
 </script>
 
 <style type="text/scss">
     @import 'styles/variables';
 
-    #context-menu {
+    @keyframes fade-in {
+        from { opacity: 0; }
+        to { opacity: 1; }
+    }
+
+    ul {
         position: absolute;
         z-index: 50;
         list-style: none;
         margin: 0;
         padding: 0;
+
+        animation: fade-in 500ms;
         background-color: white;
         box-shadow: 2px 2px 4px;
-        border-radius: 1px;
+        border-radius: $border-radius;
 
         > li {
             margin: 0;
@@ -151,34 +141,41 @@
     }
 </style>
 
-<ul
-    id="context-menu"
-    style="display: {style.display}; top: {style.top}; right: {style.right}; bottom: {style.bottom};
-    left: {style.left}"
-    on:contextmenu|stopPropagation
-    transition:fade={{ duration: 150 }}>
-    <li>
-        <a href="#what-is-here" on:click|preventDefault={handleWhatIsHere}>
-            What's here
-        </a>
-    </li>
-    <li>
-        <a href="#get-directions-from" on:click|preventDefault={handleGetDirectionsFrom}>
-            Get directions from
-        </a>
-    </li>
-    <li>
-        <a href="#get-directions-to" on:click|preventDefault={handleGetDirectionsTo}>
-            Get directions to
-        </a>
-    </li>
-    <li>
-        <a href="#center-map-here" on:click|preventDefault={handleCenterMapHere}>Center map here</a>
-    </li>
-    <li>
-        <a href="#zoom-in-here" on:click|preventDefault={handleZoomInHere}>Zoom in here</a>
-    </li>
-    <li>
-        <a href={osmUrl}>View on OpenStreetMap</a>
-    </li>
-</ul>
+{#if open}
+    <ul bind:this={element}
+        on:click={handleClick}
+        on:contextmenu|stopPropagation
+        style={style}
+    >
+        <li>
+            <a href="#what-is-here" on:click={handleWhatIsHere}>
+                What's here?
+            </a>
+        </li>
+        <li>
+            <a href="#get-directions-from" on:click={handleGetDirectionsFrom}>
+                Get directions from
+            </a>
+        </li>
+        <li>
+            <a href="#get-directions-to" on:click={handleGetDirectionsTo}>
+                Get directions to
+            </a>
+        </li>
+        <li>
+            <a href="#center-map-here"on:click={handleCenterMapHere}>
+                Center map here
+            </a>
+        </li>
+        <li>
+            <a href="#zoom-in-here" on:click={handleZoomInHere}>
+                Zoom in here
+            </a>
+        </li>
+        <li>
+            <a href={osmUrl}>
+                View on OpenStreetMap
+            </a>
+        </li>
+    </ul>
+{/if}
